delete DS_Store, *_original files
