+++ sparql/recon_reconcile_v1.sparql
@@ -6,7 +6,7 @@ PREFIX idmcore: <http://www.intavia.eu/idm-core/>
 PREFIX bd: <http://www.bigdata.com/rdf#>
 prefix bds: <http://www.bigdata.com/rdf/search#>
 
-SELECT DISTINCT ?id ?p ?name ?rank ((?score0 *100) AS ?score)
+SELECT DISTINCT ?id ?p ?name ?type ?typelabel ?pid ?rank ((?score0 *100) AS ?score)
 WHERE {
     ?name bds:search "{{q}}" .
     ?name bds:matchAllTerms "true" .
@@ -15,6 +15,14 @@ WHERE {
     ?name bds:maxRank "1000" .
     ?name bds:rank ?rank .
     ?id ?p ?name
+    {SELECT DISTINCT ?id ?type ?pid ?typelabel 
+	    WHERE {
+            ?id a ?type
+            BIND(REPLACE(str(?type), "(.*)/(.*)$", "$2") AS ?typelabel) .
+		    #?id a crm:E53_Place .
+		    #?s crm:P1_is_identified_by ?pid .
+        }
+    }
 }
 ORDER BY ?rank
 LIMIT 100
