--- sparql/recon_suggest_v1_original.sparql
@@ -1,27 +1,0 @@
-PREFIX owl: <http://www.w3.org/2002/07/owl#>
-PREFIX rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#>
-PREFIX rdfs: <http://www.w3.org/2000/01/rdf-schema#>
-PREFIX crm: <http://www.cidoc-crm.org/cidoc-crm/>
-PREFIX idmcore: <http://www.intavia.eu/idm-core/>
-PREFIX bd: <http://www.bigdata.com/rdf#>
-prefix bds: <http://www.bigdata.com/rdf/search#>
-
-SELECT DISTINCT ?id ?p ?name ?type ?pid ?score ?rank
-WHERE {
-?name bds:search "{{q}}" .
-?name bds:matchAllTerms "true" .
-?name bds:minRelevance "0.25" .
-?name bds:relevance ?score .
-?name bds:maxRank "1000" .
-?name bds:rank ?rank .
-?id ?p ?name
-{
-    SELECT DISTINCT ?id ?type ?pid
-    WHERE { 
-        ?id a ?type .
-        #?s crm:P1_is_identified_by ?pid .
-    }
-  }
-}
-ORDER BY ?rank
-LIMIT 100
